group setting is updated
